Add version of network with specifiable parameters and update platform to accomodate these additional parameters
